requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
Flask==3.0.0
gunicorn==21.2.0
idna==3.11
nba_api==1.11.3
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
selenium==4.16.0
setuptools==80.9.0
six==1.17.0
tzdata==2025.3
urllib3==2.6.2
wheel==0.45.1
